Fix code smells in CbsePdf
